add MS-DOS 16-bit (real mode) Open Watcom target
Build fed as a 16-bit DOS executable using Open Watcom's huge memory
model (-mh) with -zt=256 to keep large data items out of DGROUP.

16-bit portability fixes across the codebase:
- map int386/int386x to int86/int86x for 16-bit interrupt calls
- use far pointers (MK_FP) for video memory, BIOS data area access,
  and DOS memory block operations with _fmemcpy
- cast shift operands to long/unsigned long where shift amounts
  exceed 16 bits (undo_read_int, clipboard size, file size)
- replace 0xffff sentinel values with INT_MAX in display_line
  (0xffff truncates to -1 on 16-bit, breaking selection highlight)
- widen MYFILE.f_size to unsigned long and file_size() return to long
  so files larger than 64K can be loaded
- include <i86.h> in iowatcom.h for MK_FP and delay() prototypes

Adds CI build, packaging, and release workflow for the new target.

+++ line.c
@@ -770,7 +770,7 @@ static_inline unsigned long undo_read_int(UNDO *u)
    int w3 = u->data[--u->length];
    int w4 = u->data[--u->length];
 
-   return ((w1 << 24) | (w2 << 16) | (w3 << 8) | (w4));
+   return (((unsigned long)w1 << 24) | ((unsigned long)w2 << 16) | (w3 << 8) | w4);
 }
